[UPD] the input_redirect flow is starting to come to life. I need to now fork the process and implement the pipe from the parent process into the child process so that the child may execute the input that is given through the pipe from the parent process. I'm changing the way the environment array is passed into the execution flow so that it may be passed along with the shell structure. This gives for cleaner and more concise code.
