+++ Makefile
@@ -51,6 +51,7 @@ SRC = src/main.c \
 	src/prompt/line_edition/cursor/set_cursor.c \
 	src/redirection/redirection.c \
 	src/redirection/input_redirect.c \
+	src/redirection/tools.c \
 	src/test_helpers/print_list_test.c \
 
 OBJ = $(addsuffix .o, $(basename $(SRC)))
