+++ includes/ft_sh.h
@@ -171,7 +171,8 @@ char				*make_env_string(t_env *env_elem);
 char				**assign_redirections(char **cmd);
 t_bool				exec_redirection(t_shell *shell);
 void				input_redirect(t_shell *shell);
-
+void				do_redirect(t_shell *shell, enum e_replacement c, \
+int o_flag, void (*do_stuff_with_cmd_and_fd_now)(t_shell *, int, int));
 /*
 ** command line history functions:
 */
